logical: extract Metrics into own package
We want to start adding metrics to ldr txn mode, including
the existing metrics. Txn mode is implemented as individual
sub systems which will cause a dependency cycle if we
attempt to import the original logical package. This change
extracts the existing metrics struct to its own package.

Release note: None
