+++ pkg/crosscluster/logical/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "logical_replication_writer_processor.go",
         "lww_kv_processor.go",
         "lww_row_processor.go",
-        "metrics.go",
         "offline_initial_scan_processor.go",
         "purgatory.go",
         "resume_create_table.go",
@@ -35,6 +34,7 @@ go_library(
         "//pkg/crosscluster",
         "//pkg/crosscluster/logical/ldrdecoder",
         "//pkg/crosscluster/logical/ldrsettings",
+        "//pkg/crosscluster/logical/metrics",
         "//pkg/crosscluster/logical/sqlwriter",
         "//pkg/crosscluster/logical/txnapply",
         "//pkg/crosscluster/logical/txnmode",
@@ -115,7 +115,6 @@ go_library(
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
-        "@com_github_cockroachdb_crlib//crstrings",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_logtags//:logtags",
         "@com_github_cockroachdb_redact//:redact",
@@ -154,6 +153,7 @@ go_test(
         "//pkg/crosscluster",
         "//pkg/crosscluster/logical/ldrdecoder",
         "//pkg/crosscluster/logical/ldrtestutils",
+        "//pkg/crosscluster/logical/metrics",
         "//pkg/crosscluster/logical/sqlwriter",
         "//pkg/crosscluster/logical/txnwriter",
         "//pkg/crosscluster/replicationtestutils",
