+++ pkg/crosscluster/logical/metrics/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "metrics",
+    srcs = ["metrics.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/crosscluster/logical/metrics",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/util/metric",
+        "@com_github_cockroachdb_crlib//crstrings",
+    ],
+)
